Rename extended_rpc.rs to extended_bitcoin_rpc.rs
