+++ core/src/tx_sender/rbf.rs
@@ -861,7 +861,7 @@ pub mod tests {
     };
     use crate::constants::{MIN_TAPROOT_AMOUNT, NON_STANDARD_V3};
     use crate::errors::BridgeError;
-    use crate::extended_rpc::ExtendedBitcoinRpc;
+    use crate::extended_bitcoin_rpc::ExtendedBitcoinRpc;
     use crate::rpc::clementine::tagged_signature::SignatureId;
     use crate::rpc::clementine::{NormalSignatureKind, NumberedSignatureKind};
     use crate::task::{IntoTask, TaskExt};
